Add conversion of data type to match DB data type
With the switch to Python 3 and using the Python 3
version of the LSST stack for some prereqs, the
insertion of metadata into the database no longer
automatically converts data types.  Previously
a Boolean value read from a fits file would automatically
end up as a Number in the database.   Added code
to explicitly convert the variable's data type to
match the DB data type before calling the DB
insertion code.